Add PEPSI example spectra and shared grid helper
